Update to fix changes in pynput
Pynput updated to change the KeyCodes
pynput.keyboard.KeyCode.from_vk() to pynput.keyboard.Key.media_next
and so on
and
assert(pynput.keyboard.KeyCode.from_vk(keycode) == pynput.keyboard.Key.media_next))
fails, so added the new keys to listen to instead of the old KeyCode.from_vk()
Also added next and previous to the list of argparse arguments.
Fix not logging since I stopped it to do some debugging and forgot to
turn it on.
Make the config object clearer to read